Se refactoriza images.router.ts en dos archivos
